Move poll file to only component which uses it
